Move manifest from domain to cache
