+++ TODO.md
@@ -3,17 +3,16 @@
 ## v0.3.0
 
 1. break up domain.rs
-   1. move manifest to cache::manifest
    1. move *Spec to spiffworkflow_specs
    1. move ElementUnit* to element_units
 1. form element unit that is full workflow with call activity subprocesses removed
    1. this element unit will have the first requirement flag (LAZY_CALL_ACTIVITIES)
+1. build element units for subprocesses that map to a call activity
+   1. same as if that process was passed in with the same cache key
 1. don't always take the last element unit, let callers provide capabilities mask
 1. don't parrot `serializer_version`, return our own
    1. embed version: https://stackoverflow.com/questions/27840394/how-can-a-rust-program-access-metadata-from-its-cargo-package
 1. manifest needs to be by process id
-1. build element units for subprocesses that map to a call activity
-   1. same as if that process was passed in with the same cache key
 
 ## v0.4.0
 
@@ -21,6 +20,7 @@
 1. new api to write existing element units to the cache
 1. split cache into spiff-element-units-disk-cache
 1. split spiffworkflow-specs to own lib
+1. add support for the `all_specs` structure
 
 ## rest
