+++ module/spiff-element-units-rs/src/manifest.rs
@@ -1,4 +1,5 @@
-use crate::domain::{ElementUnits, Manifest, ManifestEntry};
+use crate::cache::manifest::{Manifest, ManifestEntry};
+use crate::domain::ElementUnits;
 
 pub fn from_element_units(element_units: &ElementUnits) -> Manifest {
     Manifest {
